CET-12088 Rename AllScorecardsPage back to the original name
