+++ src/components/ReportsPage/AllScorecardsPage/Tables/colorClasses.ts
