+++ src/components/ReportsPage/ReportsPage.tsx
@@ -23,7 +23,7 @@ import {
 } from '@backstage/core-components';
 import { Card, CardActions, CardContent, CardMedia } from '@material-ui/core';
 import { ProgressPage } from './ProgressPage';
-import { AllScorecardsPage } from './AllScorecardsPageOld/AllScorecardsPage';
+import { AllScorecardsPage } from './AllScorecardsPage/AllScorecardsPage';
 import { HeatmapPage } from './HeatmapPage';
 
 export const ReportsPageCard = ({
